==> 2_child_adult/scripts/key_cleaning_scripts/4_clean_baseline-N591.Rmd <==
---
title: "clean_baseline"
output: html_document
---

# Load packages
```{r}
library(tidyverse)
library(here)
library(naniar)
library(ggbeeswarm)
library(ggeasy)
library(readr)
library(readxl)
```

# Read in clean data

Using map to read in and bind together many csv files- notes from this blog post using option2A www.jenrichmond.rbind.io/post/use-map-to-read-many-csv-files

### read in child data
```{r message=FALSE, warning=FALSE}

data_path1 <- here("rileyraw_data", "clean_data")

files <- dir(data_path1, pattern = "*_child_*")

child_data <- files %>%
  map(~ read_csv(file.path(data_path1, .))) %>%
  reduce(rbind)



```

### read adult files
```{r message=FALSE, warning=FALSE}

data_path2 <- here("rileyraw_data", "clean_data")

files <- dir(data_path2, pattern = "*_adult_*")

adult_data <- files %>%
  map(~ read_csv(file.path(data_path2, .))) %>%
  reduce(rbind)

```

# create participant and condition columns out of filename

Use separate() to split the filename into  pp_no and condition and mutate with str_sub to drop the last 6 characters off the end of the values in the condition column to get rid of the .xlsx

```{r}
adult_cond <- adult_data %>%
  separate(file_name, into = c("pp_no", "model", "emotion"), sep = "_") %>% 
  mutate(emotion = str_sub(emotion, end = -6)) 

child_cond <- child_data %>%
  separate(file_name, into = c("pp_no", "model", "emotion"), sep = "_") %>% 
  mutate(emotion = str_sub(emotion, end = -6)) 
```


# bind child_cond and adult_cond together using rbind and then write csv
```{r}
child_adult_clean <- 
  rbind(child_cond, adult_cond) %>%
  write_csv(here("rileyraw_data", "clean_data", "child_adult_clean.csv"))

```

#read in clean data.csv

```{r}
clean_data <- read_csv(here("rileyraw_data", "clean_data", "child_adult_clean.csv"))

```


# Filter the data to only include bin 0 (baseline- BL) and plot BL scores for each trial, separately for brow and cheek, coloured by positive/negative. 

```{r}
baseline_data <- clean_data %>%
  filter(bin == "bin_0")
```
 
# make a new outlier variable 
```{r}

baseline_data <- baseline_data %>%
  group_by(pp_no, model, muscle, emotion) %>%
  mutate(meanRMS = mean(rms), sdRMS = sd(rms))

baseline_outlier <- baseline_data %>%
  mutate(outlier = if_else(rms > (meanRMS + 2*sdRMS), "TRUE", "FALSE"))

```


#make outliter logical for child and adult face data

```{r}
baseline_outlier$outlier <- as.logical(baseline_outlier$outlier)

```

#make logical variable code as 1 = true, 0 = false, so they are able to be summed for adult and child face data

```{r}
baseline_outlier_count <- baseline_outlier %>%
  group_by(pp_no, model, emotion, muscle) %>%
  summarise(out_count = sum(outlier), trial_count = n())

```

#plot outliers 

```{r}
baseline_outlier %>%
  filter(emotion == "121") %>%
  ggplot(aes(x = trial, y = rms, colour = outlier)) +
  geom_quasirandom() +
  facet_wrap(~ muscle) +
  labs(title = "condition child_121")


baseline_outlier%>%
  filter(emotion == "323") %>%
  ggplot(aes(x = trial, y = rms, colour = outlier)) +
  geom_quasirandom() +
  facet_wrap(~ muscle) +
  labs(title = "condition child_323")


baseline_outlier %>%
  filter(emotion == "424") %>%
  ggplot(aes(x = trial, y = rms, colour = outlier)) +
  geom_quasirandom() +
  facet_wrap(~ muscle) +
  labs(title = "condition child_424")


baseline_outlier %>%
  filter(emotion == "525") %>%
  ggplot(aes(x = trial, y = rms, colour = outlier)) +
  geom_quasirandom() +
  facet_wrap(~ muscle) +
  labs(title = "condition child_525")

baseline_outlier %>%
  filter(emotion == "131") %>%
  ggplot(aes(x = trial, y = rms, colour = outlier)) +
  geom_quasirandom() +
  facet_wrap(~ muscle) +
  labs(title = "condition adult_131")

baseline_outlier %>%
  filter(emotion == "232") %>%
  ggplot(aes(x = trial, y = rms, colour = outlier)) +
  geom_quasirandom() +
  facet_wrap(~ muscle) +
  labs(title = "condition adult_232")

baseline_outlier %>%
  filter(emotion == "434") %>%
  ggplot(aes(x = trial, y = rms, colour = outlier)) +
  geom_quasirandom() +
  facet_wrap(~ muscle) +
  labs(title = "condition adult_434")

baseline_outlier %>%
  filter(emotion == "535") %>%
  ggplot(aes(x = trial, y = rms, colour = outlier)) +
  geom_quasirandom() +
  facet_wrap(~ muscle) +
  labs(title = "condition adult_535")
```
